refactor: Refactored the solution and project.
