+++ HoSam.slnx
@@ -1,3 +1,3 @@
 <Solution>
-  <Project Path="HoSam.Host/HoSam.Host.csproj" />
+  <Project Path="hosam/hosam.csproj" />
 </Solution>
